+++ src/systems/physics/Physics.cc
@@ -44,6 +44,7 @@
 #include <ignition/physics/ForwardStep.hh>
 #include <ignition/physics/FrameSemantics.hh>
 #include <ignition/physics/FreeGroup.hh>
+#include <ignition/physics/FixedJoint.hh>
 #include <ignition/physics/GetContacts.hh>
 #include <ignition/physics/GetEntities.hh>
 #include <ignition/physics/Joint.hh>
@@ -82,12 +83,15 @@
 #include "ignition/gazebo/components/Geometry.hh"
 #include "ignition/gazebo/components/Gravity.hh"
 #include "ignition/gazebo/components/Inertial.hh"
+#include "ignition/gazebo/components/DetachableJoint.hh"
 #include "ignition/gazebo/components/Joint.hh"
 #include "ignition/gazebo/components/JointAxis.hh"
 #include "ignition/gazebo/components/JointPosition.hh"
+#include "ignition/gazebo/components/JointPositionReset.hh"
 #include "ignition/gazebo/components/JointType.hh"
 #include "ignition/gazebo/components/JointVelocity.hh"
 #include "ignition/gazebo/components/JointVelocityCmd.hh"
+#include "ignition/gazebo/components/JointVelocityReset.hh"
 #include "ignition/gazebo/components/LinearAcceleration.hh"
 #include "ignition/gazebo/components/LinearVelocity.hh"
 #include "ignition/gazebo/components/Link.hh"
@@ -100,6 +104,7 @@
 #include "ignition/gazebo/components/PhysicsEnginePlugin.hh"
 #include "ignition/gazebo/components/Pose.hh"
 #include "ignition/gazebo/components/PoseCmd.hh"
+#include "ignition/gazebo/components/SelfCollide.hh"
 #include "ignition/gazebo/components/Static.hh"
 #include "ignition/gazebo/components/ThreadPitch.hh"
 #include "ignition/gazebo/components/Visual.hh"
@@ -116,6 +121,9 @@ namespace components = ignition::gazebo::components;
 class ignition::gazebo::systems::PhysicsPrivate
 {
   public: using MinimumFeatureList = ignition::physics::FeatureList<
+          ignition::physics::AttachFixedJointFeature,
+          ignition::physics::DetachJointFeature,
+          ignition::physics::SetJointTransformFromParentFeature,
           // FreeGroup
           ignition::physics::FindFreeGroupFeature,
           ignition::physics::SetFreeGroupWorldPose,
@@ -388,13 +396,13 @@ void Physics::Update(const UpdateInfo &_info, EntityComponentManager &_ecm)
   if (this->dataPtr->engine)
   {
     this->dataPtr->CreatePhysicsEntities(_ecm);
+    this->dataPtr->UpdatePhysics(_ecm);
     // Only step if not paused.
     if (!_info.paused)
     {
-      this->dataPtr->UpdatePhysics(_ecm);
       this->dataPtr->Step(_info.dt);
-      this->dataPtr->UpdateSim(_ecm);
     }
+    this->dataPtr->UpdateSim(_ecm);
 
     // Entities scheduled to be removed should be removed from physics after the
     // simulation step. Otherwise, since the to-be-removed entity still shows up
@@ -470,6 +478,12 @@ void PhysicsPrivate::CreatePhysicsEntities(const EntityComponentManager &_ecm)
         if (staticComp && staticComp->Data())
         {
           model.SetStatic(staticComp->Data());
+        }
+
+        auto selfCollideComp = _ecm.Component<components::SelfCollide>(_entity);
+        if (selfCollideComp && selfCollideComp ->Data())
+        {
+          model.SetSelfCollide(selfCollideComp->Data());
         }
 
         auto modelPtrPhys = worldPtrPhys->ConstructModel(model);
@@ -667,6 +681,72 @@ void PhysicsPrivate::CreatePhysicsEntities(const EntityComponentManager &_ecm)
           _ecm.ParentEntity(_entity), false));
         return true;
       });
+
+  // Detachable joints
+  _ecm.EachNew<components::DetachableJoint>(
+      [&](const Entity &_entity,
+          const components::DetachableJoint *_jointInfo) -> bool
+      {
+        if (_jointInfo->Data().jointType != "fixed")
+        {
+          ignerr << "Detachable joint type [" << _jointInfo->Data().jointType
+                 << "] is currently not supported" << std::endl;
+          return true;
+        }
+        // Check if joint already exists
+        if (this->entityJointMap.find(_entity) != this->entityJointMap.end())
+        {
+          ignwarn << "Joint entity [" << _entity
+                  << "] marked as new, but it's already on the map."
+                  << std::endl;
+          return true;
+        }
+
+        // Check if the link entities exist in the physics engine
+        auto parentLinkPhysIt =
+            this->entityLinkMap.find(_jointInfo->Data().parentLink);
+        if (parentLinkPhysIt == this->entityLinkMap.end())
+        {
+          ignwarn << "DetachableJoint's parent link entity ["
+                  << _jointInfo->Data().parentLink << "] not found in link map."
+                  << std::endl;
+          return true;
+        }
+
+        auto childLinkPhysIt =
+            this->entityLinkMap.find(_jointInfo->Data().childLink);
+        if (childLinkPhysIt == this->entityLinkMap.end())
+        {
+          ignwarn << "DetachableJoint's child link entity ["
+                  << _jointInfo->Data().childLink << "] not found in link map."
+                  << std::endl;
+          return true;
+        }
+
+        const auto poseParent =
+            parentLinkPhysIt->second->FrameDataRelativeToWorld().pose;
+        const auto poseChild =
+            childLinkPhysIt->second->FrameDataRelativeToWorld().pose;
+
+        // Pose of child relative to parent
+        auto poseParentChild = poseParent.inverse() * poseChild;
+        auto jointPtrPhys =
+            childLinkPhysIt->second->AttachFixedJoint(parentLinkPhysIt->second);
+        if (jointPtrPhys.Valid())
+        {
+          // We let the joint be at the origin of the child link.
+          jointPtrPhys->SetTransformFromParent(poseParentChild);
+
+          igndbg << "Creating detachable joint [" << _entity << "]"
+                 << std::endl;
+          this->entityJointMap.insert(std::make_pair(_entity, jointPtrPhys));
+        }
+        else
+        {
+          ignwarn << "DetachableJoint could not be created." << std::endl;
+        }
+        return true;
+      });
 }
 
 //////////////////////////////////////////////////
@@ -716,6 +796,18 @@ void PhysicsPrivate::RemovePhysicsEntities(const EntityComponentManager &_ecm)
         }
         return true;
       });
+
+  _ecm.EachRemoved<components::DetachableJoint>(
+      [&](const Entity &_entity, const components::DetachableJoint *) -> bool
+      {
+        auto jointIt = this->entityJointMap.find(_entity);
+        if (jointIt != this->entityJointMap.end())
+        {
+          igndbg << "Detaching joint [" << _entity << "]" << std::endl;
+          jointIt->second->Detach();
+        }
+        return true;
+      });
 }
 
 //////////////////////////////////////////////////
@@ -756,7 +848,62 @@ void PhysicsPrivate::UpdatePhysics(EntityComponentManager &_ecm)
           return true;
         }
 
+        auto posReset = _ecm.Component<components::JointPositionReset>(
+            _entity);
+        auto velReset = _ecm.Component<components::JointVelocityReset>(
+            _entity);
+
+        // Reset the velocity
+        if (velReset)
+        {
+          auto& jointVelocity = velReset->Data();
+
+          if (jointVelocity.size() != jointIt->second->GetDegreesOfFreedom())
+          {
+            ignwarn << "There is a mismatch in the degrees of freedom "
+                    << "between Joint [" << _name->Data() << "(Entity="
+                    << _entity << ")] and its JointVelocityReset "
+                    << "component. The joint has "
+                    << jointIt->second->GetDegreesOfFreedom()
+                    << " while the component has "
+                    << jointVelocity.size() << ".\n";
+            }
+
+            std::size_t nDofs = std::min(
+                jointVelocity.size(), jointIt->second->GetDegreesOfFreedom());
+
+            for (std::size_t i = 0; i < nDofs; ++i)
+            {
+              jointIt->second->SetVelocity(i, jointVelocity[i]);
+            }
+        }
+
+        // Reset the position
+        if (posReset)
+        {
+          auto &jointPosition = posReset->Data();
+
+          if (jointPosition.size() != jointIt->second->GetDegreesOfFreedom())
+          {
+            ignwarn << "There is a mismatch in the degrees of freedom "
+                    << "between Joint [" << _name->Data() << "(Entity="
+                    << _entity << ")] and its JointPositionyReset "
+                    << "component. The joint has "
+                    << jointIt->second->GetDegreesOfFreedom()
+                    << " while the component has "
+                    << jointPosition.size() << ".\n";
+            }
+            std::size_t nDofs = std::min(
+                jointPosition.size(), jointIt->second->GetDegreesOfFreedom());
+            for (std::size_t i = 0; i < nDofs; ++i)
+            {
+              jointIt->second->SetPosition(i, jointPosition[i]);
+            }
+        }
+
         auto force = _ecm.Component<components::JointForceCmd>(_entity);
+        auto velCmd = _ecm.Component<components::JointVelocityCmd>(_entity);
+
         if (force)
         {
           if (force->Data().size() != jointIt->second->GetDegreesOfFreedom())
@@ -777,23 +924,42 @@ void PhysicsPrivate::UpdatePhysics(EntityComponentManager &_ecm)
         else
         {
           // Only set joint velocity if joint force is not set.
-          auto velCmd = _ecm.Component<components::JointVelocityCmd>(_entity);
+          // If both the cmd and reset components are found, cmd is ignored.
           if (velCmd)
           {
-            if (velCmd->Data().size() != jointIt->second->GetDegreesOfFreedom())
-            {
-              ignwarn << "There is a mismatch in the degrees of freedom between"
-                      << " Joint [" << _name->Data() << "(Entity=" << _entity
-                      << ")] and its JointVelocityCmd component. The joint has "
-                      << velCmd->Data().size() << " while the component has "
-                      << jointIt->second->GetDegreesOfFreedom() << ".\n";
-            }
-            std::size_t nDofs = std::min(velCmd->Data().size(),
-                jointIt->second->GetDegreesOfFreedom());
-            for (std::size_t i = 0; i < nDofs; ++i)
-            {
-              jointIt->second->SetVelocityCommand(i, velCmd->Data()[i]);
-            }
+              auto velocityCmd = velCmd->Data();
+
+              if (velReset)
+              {
+                ignwarn << "Found both JointVelocityReset and "
+                        << "JointVelocityCmd components for Joint ["
+                        << _name->Data() << "(Entity=" << _entity
+                        << "]). Ignoring JointVelocityCmd component."
+                        << std::endl;
+              }
+              else
+              {
+                if (velocityCmd.size() !=
+                      jointIt->second->GetDegreesOfFreedom())
+                {
+                  ignwarn << "There is a mismatch in the degrees of freedom"
+                          << " between Joint [" << _name->Data()
+                          << "(Entity=" << _entity<< ")] and its "
+                          << "JointVelocityCmd component. The joint has "
+                          << jointIt->second->GetDegreesOfFreedom()
+                          << " while the component has "
+                          << velocityCmd.size() << ".\n";
+                  }
+
+                  std::size_t nDofs = std::min(
+                    velocityCmd.size(),
+                    jointIt->second->GetDegreesOfFreedom());
+
+                  for (std::size_t i = 0; i < nDofs; ++i)
+                  {
+                    jointIt->second->SetVelocityCommand(i, velocityCmd[i]);
+                  }
+              }
           }
         }
 
@@ -821,14 +987,14 @@ void PhysicsPrivate::UpdatePhysics(EntityComponentManager &_ecm)
       [&](const Entity &_entity, const components::Model *,
           const components::WorldPoseCmd *_poseCmd)
       {
-        auto modelIt = this->entityModelMap.find(_entity);
-        if (modelIt == this->entityModelMap.end())
-          return true;
-
         // Get canonical link offset
         auto canonicalLink = _ecm.ChildrenByComponents(_entity,
             components::CanonicalLink());
         if (canonicalLink.empty())
+          return true;
+        auto canonicalLinkPhysIt = this->entityLinkMap.find(canonicalLink[0]);
+
+        if (canonicalLinkPhysIt == this->entityLinkMap.end())
           return true;
 
         auto canonicalPoseComp =
@@ -838,7 +1004,7 @@ void PhysicsPrivate::UpdatePhysics(EntityComponentManager &_ecm)
 
         // TODO(addisu) Store the free group instead of searching for it at
         // every iteration
-        auto freeGroup = modelIt->second->FindFreeGroup();
+        auto freeGroup = canonicalLinkPhysIt->second->FindFreeGroup();
         if (freeGroup)
         {
           freeGroup->SetWorldPose(math::eigen3::convert(_poseCmd->Data() *
@@ -853,8 +1019,8 @@ void PhysicsPrivate::UpdatePhysics(EntityComponentManager &_ecm)
           auto worldPoseComp = _ecm.Component<components::Pose>(_entity);
           if (worldPoseComp)
           {
-            auto state = worldPoseComp->SetData(_poseCmd->Data() *
-                  canonicalPoseComp->Data(), this->pose3Eql) ?
+            auto state = worldPoseComp->SetData(_poseCmd->Data(),
+                this->pose3Eql) ?
                 ComponentState::OneTimeChange :
                 ComponentState::NoChange;
             _ecm.SetChanged(_entity, components::Pose::typeId, state);
@@ -1207,6 +1373,33 @@ void PhysicsPrivate::UpdateSim(EntityComponentManager &_ecm) const
 
         return true;
       });
+
+  // Clear reset components
+  std::vector<Entity> entitiesPositionReset;
+  _ecm.Each<components::JointPositionReset>(
+      [&](const Entity &_entity, components::JointPositionReset *) -> bool
+      {
+        entitiesPositionReset.push_back(_entity);
+        return true;
+      });
+
+  for (const auto entity : entitiesPositionReset)
+  {
+    _ecm.RemoveComponent<components::JointPositionReset>(entity);
+  }
+
+  std::vector<Entity> entitiesVelocityReset;
+  _ecm.Each<components::JointVelocityReset>(
+      [&](const Entity &_entity, components::JointVelocityReset *) -> bool
+      {
+        entitiesVelocityReset.push_back(_entity);
+        return true;
+      });
+
+  for (const auto entity : entitiesVelocityReset)
+  {
+    _ecm.RemoveComponent<components::JointVelocityReset>(entity);
+  }
 
   // Clear pending commands
   _ecm.Each<components::JointForceCmd>(