+++ Changelog.md
@@ -1,5 +1,25 @@
 ## Ignition Gazebo 2.x
 
+### Ignition Gazebo 2.xx.xx (20XX-XX-XX)
+
+1.  Add DetachableJoint: A system that initially attaches two models via a fixed joint and allows for the models to get detached during simulation via a topic.
+    * [Pull Request 440](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/440)
+
+1. Update physics state even when paused (not stepping)
+    * [Pull Request 556](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/556)
+
+1. Fix entity tree context menu position
+    * [Pull Request 567](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/567)
+
+1. Fix moving static model with link offset
+    * [Pull Request 566](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/566)
+
+1. Add support for setting visual transparency through SDF
+    * [Pull Request 547](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/547)
+
+1. Add `JointPositionReset` and `JointVelocityReset` components to reset the joint state.
+    * [Pull Request 437](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/437)
+
 ### Ignition Gazebo 2.15.0 (2020-02-07)
 
 1. Fix seeking back in time in log playback